requirements.txt
Pillow
numpy
opencv-python
tqdm
